Created an experimental keyboard layout - Cyrillic with palochka.
Most languages of North Caucasus are written in the Cyrillic alphabet
with the addition of the character "palochka". This character is not
included in the common Cyrillic keyboard layouts, so the characters
'l' (Latin small L), 'I' (Latin capital I), 'І' (Ukrainian capital І)
and '1' (digit 1) are often used instead, which causes various
problems with the display and the processing of the text.
See the comment in the beginning of ext.narayam.rules.cyrl-palochka.js
for more information.

Changes in this patch set:
* resources/ext.narayam.rules.cyrl-palochka.js: the mapping itself
* Narayam.php:
  * Assigning the mapping to all the relevant languages
    that are currently supported by core MediaWiki and locally in
    translatewiki.net: Adyghe (Cyrillic), Avar, Chechen, Ingush,
    Kabardian (Cyrillic), Lak, Lezgian and Tsakhur.
  * Adding the mapping as beta.
* tests/qunit/ext.narayam.rules.tests.js: tests for the new layout.
* Narayam.i18n.php: names of the keyboard in English and Russian
  and documentation.

Patch set 2: Corrected whitespace and license according to Niklas'
             comments.

Patch set 3: rebase master.

Change-Id: I10fd1e58976e48af7a864753053a42ec84b708a8