Prevent listing of the same scheme more than once.
This is a followup to https://gerrit.wikimedia.org/r/#/c/4043/ .

That commit introduced a scheme that is used for several languages.
It appeared several times in the menu. It is not needed and it also
causes other problems.

This commit adds a check that prevents the adding of a scheme if
it has already been added.

Change-Id: I812e060b05a1b350bb3f8e2a5511a52c80d860a3