(bug 38238) Change the rule for oM and auM
'oM' always produced the OM symbol, even though often 'o' with
anusvara was needed.

In this commit 'oM' now produces o with anusvara and the OM symbol
is typed using 'AUM', according to suggestions by Siddharta and Shantanoo in the
bug comments.

The change is done for both Hindi and Marathi.

Tests were added accordingly.

Change-Id: I8cf2434e21034b7201b1074621c2712834cc1db9